requirements.txt
psycopg2-binary==2.9.9
docker==7.0.0
colorama==0.4.6 